Added: x64 windows build

+++ src/game-init.h
@@ -55,7 +55,7 @@ class game_init_box_t : public gui_object_t {
   void draw_box_string(int x, int y, const std::string &str);
   void draw_player_box(int player, int x, int y);
   void draw_backgraund();
-  int get_player_face_sprite(int face);
+  unsigned int get_player_face_sprite(size_t face);
   void handle_action(int action);
   bool handle_player_click(int player, int x, int y);
   void apply_random(random_state_t rnd);
